replaced Sign-Up file in Components folder
